+++ app/controllers/topics_controller.rb
@@ -1,4 +1,5 @@
 class TopicsController < ApplicationController
+  before_action :authenticate!, only: [:create, :edit, :update, :new, :destroy]
 
   def index
     @topics = Topic.all
@@ -6,36 +7,47 @@ class TopicsController < ApplicationController
 
   def new
     @topic= Topic.new
+    authorize @topic
   end
 
   def create
-    @topic = Topic.new(topic_params)
+    @topic = current_user.topics.build(topic_params)
+    authorize @topic
 
     if @topic.save
+      flash[:success] = "You've created a new topic."
       redirect_to topics_path
     else
-      render new_topic_path
+      flash[:danger] = @topic.errors.full_messages
+      redirect_to new_topic_path
     end
 
   end
 
   def edit
     @topic = Topic.find_by(id:params[:id])
+    authorize @topic
   end
 
   def update
     @topic = Topic.find_by(id: params[:id])
+    authorize @topic
 
     if @topic.update(topic_params)
+      flash[:success] = "You've updated the topic."
       redirect_to topics_path
     else
+      flash[:danger] = @topic.errors.full_messages
       redirect_to edit_topic_path(@topic)
     end
   end
 
   def destroy
     @topic = Topic.find_by(id: params[:id])
+    authorize @topic
+
     if @topic.destroy
+      flash[:success] = "You've deleted the topic."
       redirect_to topics_path
     else
       redirect_to topic_path(@topic)
